Fixing hypermedia generation & tests
