+++ src/utils/hypermedia.js
@@ -1,15 +1,6 @@
 import {URL} from "url";
 import {keysort} from "keysort";
-import {
-	collection as collectionPattern,
-	hypermedia as hypermediaPattern,
-	trailing,
-	trailingS,
-	trailingSlash,
-	trailingY
-} from "./regex";
-import {id} from "./id";
-import {scheme} from "./scheme";
+import {collection as collectionPattern, trailingSlash} from "./regex";
 import {
 	COLLECTION,
 	COMMA_SPACE,
@@ -19,7 +10,6 @@ import {
 	FIRST,
 	GET,
 	HEADER_SPLIT,
-	IE,
 	INT_0,
 	INT_1,
 	INT_200,
@@ -33,13 +23,11 @@ import {
 	PAGE_SIZE,
 	PREV,
 	REL_URI,
-	RELATED,
-	S,
 	SLASH,
 	URL_127001
 } from "../core/constants.js";
+import {marshal} from "./marshal.js";
 
-// @todo audit this function
 export function hypermedia (req, res, rep) {
 	const server = req.server,
 		headers = res.getHeaders(),
@@ -48,53 +36,6 @@ export function hypermedia (req, res, rep) {
 		seen = new Set(),
 		exists = rep !== null;
 	let query, page, page_size, nth, root, parent;
-
-	// Parsing the object for hypermedia properties
-	function marshal (obj, rel, item_collection) {
-		let keys = Object.keys(obj),
-			lrel = rel || RELATED,
-			result;
-
-		if (keys.length === INT_0) {
-			result = null;
-		} else {
-			for (const i of keys) {
-				if (obj[i] !== void 0 && obj[i] !== null) {
-					const lid = id(i);
-					let lcollection, uri;
-
-					// If ID like keys are found, and are not URIs, they are assumed to be root collections
-					if (lid || hypermediaPattern.test(i)) {
-						const lkey = obj[i].toString();
-
-						if (lid === false) {
-							lcollection = i.replace(trailing, EMPTY).replace(trailingS, EMPTY).replace(trailingY, IE) + S;
-							lrel = RELATED;
-						} else {
-							lcollection = item_collection;
-							lrel = ITEM;
-						}
-
-						if (scheme(lkey) === false) {
-							uri = `${lcollection[0] === SLASH ? EMPTY : SLASH}${lcollection.replace(/\s/g, ENCODED_SPACE)}/${lkey.replace(/\s/g, ENCODED_SPACE)}`;
-
-							if (uri !== root && seen.has(uri) === false) {
-								seen.add(uri);
-
-								if (server.allowed(GET, uri)) {
-									links.push({uri: uri, rel: lrel});
-								}
-							}
-						}
-					}
-				}
-			}
-
-			result = obj;
-		}
-
-		return result;
-	}
 
 	query = req.parsed.searchParams;
 	page = Number(query.get(PAGE)) || INT_1;
@@ -124,10 +65,6 @@ export function hypermedia (req, res, rep) {
 	if (exists) {
 		if (Array.isArray(rep.data)) {
 			if (req.method === GET && (rep.status >= INT_200 && rep.status <= INT_206)) {
-				if (isNaN(page) || page <= INT_0) {
-					page = INT_1;
-				}
-
 				nth = Math.ceil(rep.data.length / page_size);
 
 				if (nth > INT_1) {
@@ -163,14 +100,14 @@ export function hypermedia (req, res, rep) {
 			if (req.hypermedia) {
 				for (const i of rep.data) {
 					if (i instanceof Object) {
-						marshal(i, ITEM, req.parsed.pathname.replace(trailingSlash, EMPTY));
+						marshal(i, ITEM, req.parsed.pathname.replace(trailingSlash, EMPTY), root, seen, links, server);
 					} else {
 						const li = i.toString();
 
 						if (li !== collection) {
 							const uri = li.indexOf(DOUBLE_SLASH) >= INT_0 ? li : `${collection.replace(/\s/g, ENCODED_SPACE)}/${li.replace(/\s/g, ENCODED_SPACE)}`.replace(/^\/\//, SLASH);
 
-							if (server.allowed(GET, uri)) {
+							if (uri !== collection && server.allowed(GET, uri)) {
 								links.push({uri: uri, rel: ITEM});
 							}
 						}
@@ -184,7 +121,7 @@ export function hypermedia (req, res, rep) {
 				parent.pop();
 			}
 
-			rep.data = marshal(rep.data, void 0, parent[parent.length - INT_1]);
+			rep.data = marshal(rep.data, void 0, parent[parent.length - INT_1], root, seen, links, server);
 		}
 	}
 
@@ -200,7 +137,7 @@ export function hypermedia (req, res, rep) {
 
 		res.header(LINK, keysort(links, REL_URI).map(i => `<${i.uri}>; rel="${i.rel}"`).join(COMMA_SPACE));
 
-		if (exists && rep.links !== void 0) {
+		if (exists && Array.isArray(rep?.links ?? EMPTY)) {
 			rep.links = links;
 		}
 	}
